+++ internal/mwdd/files/embed/mediawiki/MwddSettings.php
@@ -126,6 +126,7 @@ if ( $dashboardServicesCache !== null ) {
 		'shellbox-syntaxhighlight' => ( $dashboardServicesCache['shellbox-syntaxhighlight'] ?? false ) && !defined( 'MW_PHPUNIT_TEST' ),
 		'shellbox-timeline' => ( $dashboardServicesCache['shellbox-timeline'] ?? false ) && !defined( 'MW_PHPUNIT_TEST' ),
 		'prometheus' => ( $dashboardServicesCache['prometheus'] ?? false ) && !defined( 'MW_PHPUNIT_TEST' ),
+		'grafana' => ( $dashboardServicesCache['grafana'] ?? false ) && !defined( 'MW_PHPUNIT_TEST' ),
 	];
 } else {
 	// Fallback to direct DNS checks (slow, but works if dashboard is not running)
@@ -150,6 +151,7 @@ if ( $dashboardServicesCache !== null ) {
 		'shellbox-syntaxhighlight' => gethostbyname('shellbox-syntaxhighlight') !== 'shellbox-syntaxhighlight' && !defined( 'MW_PHPUNIT_TEST' ),
 		'shellbox-timeline' => gethostbyname('shellbox-timeline') !== 'shellbox-timeline' && !defined( 'MW_PHPUNIT_TEST' ),
 		'prometheus' => gethostbyname('prometheus') !== 'prometheus' && !defined( 'MW_PHPUNIT_TEST' ),
+		'grafana' => gethostbyname('grafana') !== 'grafana' && !defined( 'MW_PHPUNIT_TEST' ),
 	];
 }
 // $serviceLookupEndTime = microtime( true );